A first attempt at wrapping Riverine in xarray simlab
This is an initial approach for wrapping Sarah's clearwater riverine into a xarray simlab process. The model does appear to execute for the specified number of timesteps, but I have not yet had success in getting the mesh variable to output (at least on model completion). That will be a critical step for data exchange between timesteps.